Fix all inputs containing non-ASCII characters got garbled
As of network-2.4.1.0, Network.URI.unEscapeString converts an input to a Unicode string assuming the input is UTF-8 encoded bytes, as well as decodes the percent-encoding. As a result, this library converts all inputs to Unicode strings, then truncates all characters in the strings to 8bits with Data.ByteString.Lazy.Char8.pack.

I copied unEscapeString from the old network library that just unescapes the percent-encoding.