Run doctests in CI rather than via Cabal.
Running doctests in a reliable and portable fashion is very hard, because the
doctest suite needs to know many of the compiler flags "cabal configure"
determined for the build to evaluate the code successfully. This can be
accomplished by writing a custom build system that's based on top of the
package http://hackage.haskell.org/package/cabal-doctest. Alternatively, modern
versions of cabal-install offer a "cabal doctest" function that can be used to
run tests in the available code.

Most people opt to not register doctests as a test suite instead, but to simply
execute "cabal doctest" in their travis-ci script to ensure the comments always
match the actual code. This seems much easier for us, too, especially
considering that our entire code base has just a single doctest comment.