Update code for network 3.0.x.
We dropped the obsolete Network.CGI.Compat module. The code in that module
relied on obsolete functions and types from `network` which have been dropped
there in the latest 3.x release. Re-writing the Compat module to use the new
types felt like it would defeat the purose of the module, so we've dropped it
instead.